+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [Issue #2926] Fix problem with escaping characters on creation.
 - [Issue #2932] Show username and user photo in slack messages.
 - Fixed problem with test button.
+- Allow to configure what to notify.
 
 
 ## 0.0.8 (2015-06-18)